+++ src/docs/lotto_requirements.md
@@ -11,5 +11,22 @@
   - 핵심 로직을 구현하는 코드와 UI를 담당하는 로직을 구분한다.
   - UI 로직을 InputView, ResultView와 같은 클래스를 추가해 분리한다.
 
-### 기능 목록 정리
-- 기능을 구현하기 전에 README.md 파일에 구현할 기능 목록을 정리해 추가한다.
+## ✨ 기능 목록 정리 
+- **기능을 구현하기 전에 README.md 파일에 구현할 기능 목록을 정리해 추가한다.**
+1. 구입 금액을 입력 받아야한다.
+   1. [예외] 1000의 배수가 아닌경우.
+
+
+2. 구입 금액 만큼 로또가 생성된다.
+
+
+3. 생성된 로또의 번호를 출력한다.
+   1. 로또의 번호는 오름차순으로 정렬 한다.
+
+
+4. 당첨 번호를 입력 받아야한다.
+   1. [예외 ]당첨 번호가 6자리가 아닌경우.
+
+
+5. 당첨 통계를 출력한다.
+   1. 당첨 통계는 3개 일치 ~ 6개 일치가 각 몇개씩 일치하는지 출력한다.
