+++ src/main/java/lotto/LottoApplication.java
@@ -1,9 +1,10 @@
 package lotto;
 
+import lotto.domain.LottoMachine;
 import lotto.view.InputView;
 import lotto.view.ResultView;
 
-public class Application {
+public class LottoApplication {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         int money = inputView.askPurchaseAmount();
