+++ src/main/java/lotto/domain/Lotties.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.*;
 
@@ -22,5 +22,13 @@ public class Lotties {
             lotties.get(i).getLottoRank(winningNumber);
         }
     }
+    /*
+    public LottoStatistics matchResult(Lotto winningLotto) {
+        // this.lotties를 순환하면서 winningLotto와 비교하여 Rank 계산
+        // -> Rank의 개수를 담은 resultMap 생성
+        return new LottoStatistics(resultMap);
+    }
+     */
+
 
 }
