+++ src/main/java/lotto/domain/RandomNumGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.*;
