+++ src/main/java/lotto/view/ResultView.java
@@ -1,12 +1,12 @@
 package lotto.view;
 
-import lotto.Lotto;
-import lotto.LottoMachine;
-import lotto.SameNumbersCount;
+import lotto.domain.Lotto;
+import lotto.domain.LottoMachine;
+import lotto.domain.SameNumbersCount;
 
 public class ResultView {
     private static final String PURCHASE_MESSAGE = "%d개를 구매했습니다.\n";
-    private static final String WINNING_STATISTICS = "당첨 통계\n";
+    private static final String WINNING_STATISTICS = "당첨 통계";
     private static final String LINE = "-------\n";
 
     public void printLotties(LottoMachine lottoMachine) {
@@ -19,7 +19,7 @@ public class ResultView {
     }
 
     public void printSameNumbers() {
-        System.out.print(WINNING_STATISTICS);
+        System.out.println(WINNING_STATISTICS);
         System.out.print(LINE);
 
         for (SameNumbersCount s : SameNumbersCount.values()) {
