+++ src/test/java/lotto/RandomNumGeneratorTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.RandomNumGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
