+++ src/test/java/lotto/StringUtilTest.java
@@ -1,10 +1,10 @@
 package lotto;
 
+import lotto.util.StringUtil;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class StringUtilTest {
